Try to recognize M02S and print
This is a quick and dirty hack to recognize and print to the M02S thermal printer from a Mac OS desktop, rather than be forced to use the Phomemo app. When dragging an image in, it should be about 1:1.9 aspect ratio or 512 x 270 @ 300 DPI. It'd be cool to be able to define custom PRINTER_PHOMEMO settings, but I don't know enough about BLE comms to get into the weeds.